requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
requests-html==0.10.0
beautifulsoup4==4.12.2
websockets==12.0
python-dotenv==1.0.0
lxml==4.9.3
cloudscraper==1.2.71
